pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "maps-scraper"
version = "2.0.0"
description = "Google Maps business contact scraper with email enrichment"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
keywords = ["google-maps", "scraper", "lead-generation", "playwright"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "playwright>=1.44.0,<2.0",
    "requests>=2.31.0,<3.0",
    "PyYAML>=6.0.1,<7.0",
    "urllib3>=2.0.0,<3.0",
]

[project.optional-dependencies]
excel = ["openpyxl>=3.1.0"]
dev   = ["pytest>=8.2.0", "pytest-cov>=5.0.0"]

[project.scripts]
maps-scraper = "maps_scraper:main"

[project.urls]
# TODO: update this URL before publishing
Homepage   = "https://github.com/your-github-username/google-maps-scraper"
# TODO: update this URL before publishing
Repository = "https://github.com/your-github-username/google-maps-scraper"
# TODO: update this URL before publishing
Issues     = "https://github.com/your-github-username/google-maps-scraper/issues"
Changelog  = "https://github.com/yourusername/google-maps-scraper/blob/main/CHANGELOG.md"

[tool.setuptools.packages.find]
where = ["."]
include = ["scraper*"]

[tool.pytest.ini_options]
testpaths    = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["scraper", "maps_scraper"]
omit   = ["tests/*", "*/__pycache__/*"]

[tool.coverage.report]
show_missing = true
skip_covered = false
